More options for directives
Some additional optional ways directives can be handled. By using an
alias on the @Input, can then use [] around the alias name and assign
the value. If the input is a string value, can omit the [] and assign
directly to a string, as long as don't include the single quotes on the
string.